GOT THE MUX api uploader working in the html!!!!
* Both the inline html Mux uploader works AND the net::HTTP version
work; the mux inline html version is nice as it updates live and is less
buggy, the downside is that we do not update a database with additional
information for editing purposes/revisions purposes. The hacked
net::HTTP version is working but not updated syncrhnously, and we will
need to extract the playback id and update a database as we go along.