Informal logging, Separate Upload Video
Prepare to iterate over assets stored in database, instead of through
MUX script, to display the video list.

The reason is to make it easier for us to decide what to store aside the
assets.

If we extract the playback ID and store it any other meta-data chosen,
such as User uploaded, or normal style IDv3 information.

We can use the IMDB API alongside the AWS account, and gain more
information and details about the video asset.