WIP:Add thumbnail attribute
* The thumbnail attribute stores the precise second which should be displayed
and viewable before the video has been clicked on and started.

* This would then get passed to Mux in the front end:

```
     <mux-player
            playback-id= <%= asset.playback_id%>
            metadata-video-title="Placeholder (optional)"
            metadata-viewer-user-id="Placeholder (optional)"
            primary-color="#d38a8a"
            accent-color="#d38a8a"
            thumbnail-time="107" <--------------------HERE
            title=" <%= asset.title %>"
        ></mux-player>
```
* Although the thumbnail-time is registered as a string, I believe it
should be stored as an integer. My reasoning for this is:

The value that is currently visible in the above given example (107
seconds) is 1 minute and 47 seconds. Whilst this may be ok for shorter
videos, calculating the seconds in larger time intervals will quickly
become a nuissance, and so there will have to be some operation to
translate given hours, minutes and seconds into seconds only. I am not
certain if the thumbnail_position/time is stored as seconds, or
hours,minutes, seconds, like this: 1.30.2 - though that is not an
integer and therefore would not work... 1:30:02 - nope. Ok. the
calculation will be done before, and the value stored in seconds like
so: 147