Styles: arrange admin upload forms
* Upload functionality is now at the bottom of the page,
it should be at the bottom of the page, not at the top, because it
follows the line of site that the user goes through, but a message
should be included to signal to behaviour of the form (i.e. once a video
is uploaded, it can ship - with or without metadata).
* Make descriptino box bigger to encourage more writing on behalf of the
user.